clear up some compiler warnings (non-generics)
git-svn-id: https://svn.apache.org/repos/asf/lucene/dev/branches/branch_3x@1004086 13f79535-47bb-0310-9956-ffa450edef68

+++ solr/src/java/org/apache/solr/search/function/ByteFieldSource.java
@@ -53,7 +53,7 @@ public class ByteFieldSource extends FieldCacheSource {
     return new DocValues() {
       @Override
       public byte byteVal(int doc) {
-        return (byte) arr[doc];
+        return arr[doc];
       }
 
       @Override
